+++ py_gql/sdl/__init__.py
@@ -3,6 +3,7 @@
 Utilities to work with the GraphQL schema definition language (SDL).
 """
 
+from .ast_schema_printer import ASTSchemaPrinter
 from .schema_directives import SchemaDirective, apply_schema_directives
 from .schema_from_ast import build_schema, extend_schema
 
@@ -11,4 +12,5 @@ __all__ = (
     "extend_schema",
     "SchemaDirective",
     "apply_schema_directives",
+    "ASTSchemaPrinter",
 )